requirements.txt
discord.py==1.7.1
dislash.py==1.0.13
boto==2.49.0